requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
pandas>=2.0,<3.0
numpy>=1.24,<3.0
openpyxl>=3.1,<4.0
python-dotenv>=1.0,<2.0
